# hirgo/openapi.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"security":[{"bearerAuth":[]}],"tags":[{"name":"Lookup Management","description":"APIs to retrieve lookup values for different types"},{"name":"Company Management","description":"Endpoints for managing companies"}],"paths":{"/api/v1/vacancies/{id}":{"get":{"tags":["vacancy-controller"],"summary":"Get vacancy by ID","description":"Retrieves a vacancy by its ID.","operationId":"getVacancyById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Vacancy found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/VacancyResponse"}}}},"404":{"description":"Vacancy not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/VacancyResponse"}}}}}},"put":{"tags":["vacancy-controller"],"summary":"Update a vacancy","description":"Updates an existing vacancy by its ID.","operationId":"updateVacancy","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateVacancyRequest"}}},"required":true},"responses":{"200":{"description":"Vacancy updated successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/VacancyResponse"}}}},"404":{"description":"Vacancy not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/VacancyResponse"}}}},"400":{"description":"Invalid input","content":{"*/*":{"schema":{"$ref":"#/components/schemas/VacancyResponse"}}}},"403":{"description":"Unauthorized - you don't own this vacancy","content":{"*/*":{"schema":{"$ref":"#/components/schemas/VacancyResponse"}}}}}},"delete":{"tags":["vacancy-controller"],"summary":"Delete a vacancy","description":"Deletes a vacancy by its ID.","operationId":"deleteVacancy","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"204":{"description":"Vacancy deleted successfully"},"404":{"description":"Vacancy not found"},"403":{"description":"Unauthorized - you don't own this vacancy"}}}},"/api/v1/companies/{id}":{"get":{"tags":["Company Management"],"summary":"Get Company by ID","description":"Retrieve company details by its ID","operationId":"getCompanyById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CompanyResponse"}}}}}},"put":{"tags":["Company Management"],"summary":"Update a Company","description":"Update company details by ID","operationId":"updateCompany","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCompanyRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CompanyResponse"}}}}},"security":[{"bearerAuth":[]}]},"delete":{"tags":["Company Management"],"summary":"Delete a Company","description":"Delete a company by ID","operationId":"deleteCompany","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}},"security":[{"bearerAuth":[]}]}},"/api/v1/vacancies":{"get":{"tags":["vacancy-controller"],"summary":"Get all vacancies","description":"Retrieves a list of all vacancies for public display.","operationId":"getAllVacancies","responses":{"200":{"description":"List of vacancies retrieved successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/VacancyListResponse"}}}}}},"post":{"tags":["vacancy-controller"],"summary":"Create a new vacancy","description":"Creates a new vacancy and returns the created vacancy details.","operationId":"createVacancy","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateVacancyRequest"}}},"required":true},"responses":{"201":{"description":"Vacancy created successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/VacancyCreateResponse"}}}},"400":{"description":"Invalid input","content":{"*/*":{"schema":{"$ref":"#/components/schemas/VacancyCreateResponse"}}}}}}},"/api/v1/vacancies/filter":{"post":{"tags":["vacancy-controller"],"summary":"Filter vacancies","description":"Filters vacancies based on the provided criteria.","operationId":"filterVacancies","parameters":[{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VacancyFilterRequest"}}},"required":true},"responses":{"200":{"description":"Filtered vacancies retrieved successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PagedModel"}}}},"400":{"description":"Invalid filter criteria","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PagedModel"}}}}}}},"/api/v1/companies":{"get":{"tags":["Company Management"],"summary":"Get All Companies","description":"Retrieve a list of all companies","operationId":"getAllCompanies","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CompanyResponse"}}}}}}},"post":{"tags":["Company Management"],"summary":"Create a Company","description":"Create a new company","operationId":"createCompany","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCompanyRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CompanyResponse"}}}}},"security":[{"bearerAuth":[]}]}},"/api/auth/register/company":{"post":{"tags":["registration-controller"],"operationId":"registerCompany","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CompanyRegisterRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}}}}},"/api/auth/register/applicant":{"post":{"tags":["registration-controller"],"operationId":"registerApplicant","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApplicantRegisterRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}}}}},"/api/auth/refresh-token":{"post":{"tags":["auth-controller"],"operationId":"refreshToken","parameters":[{"name":"refreshToken","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}}}}},"/api/auth/logout":{"post":{"tags":["auth-controller"],"operationId":"logout","parameters":[{"name":"token","in":"query","required":false,"schema":{"type":"string"}},{"name":"refreshToken","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/auth/login":{"post":{"tags":["auth-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}}}}},"/api/auth/introspect":{"post":{"tags":["auth-controller"],"operationId":"introspectToken","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/IntrospectionResponse"}}}}}}},"/api/v1/vacancies/company":{"get":{"tags":["vacancy-controller"],"summary":"Get company vacancies","description":"Retrieves all vacancies belonging to the authenticated company.","operationId":"getCompanyVacancies","responses":{"200":{"description":"Company's vacancies retrieved successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/VacancyListResponse"}}}},"403":{"description":"Access denied - requires company role","content":{"*/*":{"schema":{"$ref":"#/components/schemas/VacancyListResponse"}}}}}}},"/api/v1/lookups":{"get":{"tags":["Lookup Management"],"summary":"Retrieve a lookup by ID","description":"Fetches a specific lookup value based on its ID. Optionally filters by type.","operationId":"getLookup","parameters":[{"name":"id","in":"query","description":"ID of the lookup value to retrieve","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"type","in":"query","description":"Optional filter for lookup type (EducationLevel, EmploymentType, ExperienceLevel, LanguageSkills, Location, WorkSchedule)","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LookupResponse"}}}}}}},"/api/v1/lookups/all":{"get":{"tags":["Lookup Management"],"summary":"Retrieve all lookup values of a specific type","description":"Fetches all lookup values belonging to a given type. Available types: EducationLevel, EmploymentType, ExperienceLevel, LanguageSkills, Location, WorkSchedule.","operationId":"getLookups","parameters":[{"name":"type","in":"query","description":"Type of lookup values to retrieve (EducationLevel, EmploymentType, ExperienceLevel, LanguageSkills, Location, WorkSchedule)","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/LookupResponse"}}}}}}}},"/api/v1/companies/search":{"get":{"tags":["Company Management"],"summary":"Search Companies","description":"Search companies by name","operationId":"searchCompaniesByName","parameters":[{"name":"name","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CompanyResponse"}}}}}}}},"/api/v1/categories":{"get":{"tags":["category-controller"],"operationId":"getAllCategories","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CategoryListResponse"}}}}}}},"/api/v1/categories/{id}":{"get":{"tags":["category-controller"],"operationId":"getCategoryById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CategoryResponse"}}}}}}}},"components":{"schemas":{"Description":{"type":"object","properties":{"responsibilities":{"type":"array","items":{"type":"string"}},"education":{"type":"array","items":{"type":"string"}},"experience":{"type":"array","items":{"type":"string"}},"requiredSkills":{"type":"array","items":{"type":"string"}},"preferredSkills":{"type":"array","items":{"type":"string"}}}},"UpdateVacancyRequest":{"type":"object","properties":{"title":{"type":"string"},"experienceLevel":{"type":"string","enum":["INTERN","JUNIOR","MIDDLE","SENIOR","LEAD"]},"locationType":{"type":"string","enum":["BAKU","GANJA","SUMGAIT","MINGACHEVIR","SHAMAKHI","GUBA","SHAKI","LANKARAN","NAKHCHIVAN","GAZAKH","AGDAM","AGDASH","AGSU","ASTARA","BALAKAN","BARDAA","BEYLQAN","BILASUVAR","DASHKESAN","FIZULI","GADABEY","GORANBOY","HACIQABUL","IMISHLI","ISMAYILLI","KURDAMIR","LERIK","MASALLI","YEVLAKH","ZARDAB","SABIRABAD","SADARAK","SAATLI","SABRAN","SIAZAN","ZAQATALA","ZANGILAN"]},"salary":{"type":"integer","format":"int64"},"applicationDeadline":{"type":"string","format":"date-time"},"categoryId":{"type":"string"},"workSchedule":{"type":"string","enum":["FULL_TIME","PART_TIME","FREE_LANCE","REMOTE"]},"description":{"$ref":"#/components/schemas/Description"}}},"VacancyResponse":{"type":"object","properties":{"id":{"type":"string"},"title":{"type":"string"},"companyName":{"type":"string"},"companyLogo":{"type":"string"},"postedAt":{"type":"string","format":"date-time"},"experienceLevelId":{"type":"integer","format":"int64"},"locationTypeId":{"type":"integer","format":"int64"},"salary":{"type":"integer","format":"int64"},"applicationDeadline":{"type":"string","format":"date-time"},"categoryId":{"type":"integer","format":"int64"},"languageSkillsIds":{"type":"array","items":{"type":"integer","format":"int64"}},"workScheduleId":{"type":"integer","format":"int64"},"description":{"$ref":"#/components/schemas/Description"}}},"UpdateCompanyRequest":{"type":"object","properties":{"name":{"type":"string"},"logo":{"type":"string"},"location":{"type":"string"},"phoneNumber":{"type":"string"},"address":{"type":"string"}}},"CompanyResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"description":{"type":"string"},"location":{"type":"string"},"phoneNumber":{"type":"string"},"address":{"type":"string"}}},"CreateVacancyRequest":{"type":"object","properties":{"title":{"type":"string"},"experienceLevelId":{"type":"integer","format":"int64"},"locationTypeId":{"type":"integer","format":"int64"},"employmentTypeId":{"type":"integer","format":"int64"},"educationLevelId":{"type":"integer","format":"int64"},"languageSkillsIds":{"type":"array","items":{"type":"integer","format":"int64"}},"workScheduleId":{"type":"integer","format":"int64"},"salary":{"type":"integer","format":"int64"},"applicationDeadline":{"type":"string","format":"date-time"},"categoryId":{"type":"integer","format":"int64"},"description":{"$ref":"#/components/schemas/VacancyDescription"}}},"VacancyDescription":{"type":"object","properties":{"responsibilities":{"type":"array","items":{"type":"string"}},"education":{"type":"array","items":{"type":"string"}},"experience":{"type":"array","items":{"type":"string"}},"requiredSkills":{"type":"array","items":{"type":"string"}},"preferredSkills":{"type":"array","items":{"type":"string"}}}},"VacancyCreateResponse":{"type":"object","properties":{"id":{"type":"string"},"title":{"type":"string"},"postedAt":{"type":"string","format":"date-time"},"experienceLevelId":{"type":"integer","format":"int64"},"locationTypeId":{"type":"integer","format":"int64"},"salary":{"type":"integer","format":"int64"},"applicationDeadline":{"type":"string","format":"date-time"},"categoryId":{"type":"integer","format":"int64"},"languageSkillsIds":{"type":"array","items":{"type":"integer","format":"int64"}},"workScheduleId":{"type":"integer","format":"int64"},"description":{"$ref":"#/components/schemas/Description"}}},"VacancyFilterRequest":{"type":"object","properties":{"locationTypeId":{"type":"integer","format":"int64"},"categoryId":{"type":"integer","format":"int64"},"employmentTypeId":{"type":"integer","format":"int64"},"workScheduleId":{"type":"integer","format":"int64"},"minSalary":{"type":"integer","format":"int64"},"maxSalary":{"type":"integer","format":"int64"},"postedAfter":{"type":"string","format":"date-time"},"postedBefore":{"type":"string","format":"date-time"}}},"Pageable":{"type":"object","properties":{"page":{"type":"integer","format":"int32","minimum":0},"size":{"type":"integer","format":"int32","minimum":1},"sort":{"type":"array","items":{"type":"string"}}}},"PageMetadata":{"type":"object","properties":{"size":{"type":"integer","format":"int64"},"number":{"type":"integer","format":"int64"},"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int64"}}},"PagedModel":{"type":"object","properties":{"content":{"type":"array","items":{"type":"object"}},"page":{"$ref":"#/components/schemas/PageMetadata"}}},"CreateCompanyRequest":{"type":"object","properties":{"name":{"type":"string"},"logo":{"type":"string"},"location":{"type":"string"},"phoneNumber":{"type":"string"},"address":{"type":"string"}}},"CompanyRegisterRequest":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"},"name":{"type":"string"},"phone":{"type":"string"}},"required":["email","name","password","phone"]},"TokenResponse":{"type":"object","properties":{"accessToken":{"type":"string"},"refreshToken":{"type":"string"},"idToken":{"type":"string"}}},"ApplicantRegisterRequest":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"},"dateOfBirth":{"type":"string","format":"date"},"phone":{"type":"string"}},"required":["dateOfBirth","email","firstName","lastName","password","phone"]},"LoginRequest":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"IntrospectionResponse":{"type":"object","properties":{"active":{"type":"boolean"},"username":{"type":"string"},"tokenType":{"type":"string"},"expiresAt":{"type":"integer","format":"int64"}}},"VacancyListResponse":{"type":"object","properties":{"vacancies":{"type":"array","items":{"$ref":"#/components/schemas/VacancyResponse"}}}},"LookupResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"type":{"type":"string"}}},"CategoryListResponse":{"type":"object","properties":{"categories":{"type":"array","items":{"$ref":"#/components/schemas/CategoryResponse"}}}},"CategoryResponse":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string","enum":["IT","MARKETING","EDUCATION","FINANCE","HEALTHCARE","ENGINEERING","SALES_AND_MARKETING","DESIGN_AND_CREATIVITY","LAW","TOURISM_AND_HOSPITALITY","ADMINISTRATIVE_JOBS","CONSTRUCTION_AND_ARCHITECTURE","TRANSPORT_AND_LOGISTICS","CUSTOMER_SERVICE","MEDIA_AND_COMMUNICATION"]}}}},"securitySchemes":{"bearerAuth":{"type":"http","name":"bearerAuth","scheme":"bearer","bearerFormat":"JWT"}}}}